Feat/Refactor the workout form
