+++ astra_web/src/app/main/layout.tsx
@@ -1,5 +1,4 @@
 "use client";
-import ThemeToggle from "@/components/theme-toggle";
 import type { ReactNode } from "react";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
 import { AppSidebar } from "@/components/app-sidebar";
@@ -10,7 +9,6 @@ export default function Layout({ children }: { children: ReactNode }) {
       <AppSidebar />
       <div className="min-h-screen w-full bg-gray-50 dark:bg-gray-900 p-6">
         <SidebarTrigger />
-        <ThemeToggle />
         {children}
       </div>
     </SidebarProvider>
