--- astra_web/src/services/fitness/domain/create-fitness-goal-dto.ts
@@ -1,7 +1,0 @@
-export interface CreateFitnessGoalDTO {
-  category: string;
-  description: string | null;
-  targetDate: Date;
-  targetValue: number;
-  title: string;
-}
