--- astra_web/src/services/fitness/domain/daily-activity.ts
@@ -1,4 +1,0 @@
-export interface DailyActivity {
-  date: string;
-  caloriesBurnt: number;
-}
