--- astra_web/src/services/fitness/domain/fitness-goal.ts
@@ -1,9 +1,0 @@
-export interface FitnessGoal {
-  category: string;
-  currentValue: number;
-  description: string | null;
-  targetDate: Date;
-  targetValue: number;
-  title: string;
-  unit: string;
-}
