+++ astra_web/src/services/fitness/domain/get-workout-vo.ts
@@ -1,0 +1,7 @@
+import type { Exercise, Run, Workout } from "./workout";
+
+export interface GetWorkoutVO {
+  exercises: Exercise[];
+  runs: Run[];
+  workout: Workout;
+}
