+++ astra_web/src/services/fitness/domain/get-workouts-vo.ts
@@ -1,0 +1,6 @@
+import { Workout } from "./workout";
+
+export interface GetWorkoutsVO {
+  workouts: Workout[];
+  total: number;
+}
