+++ backend/astra_app/src/main/java/com/wenjun/astra_app/model/vo/GetWorkoutVO.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Builder
 @Data
 public class GetWorkoutVO {
-    private WorkoutLogEntity workoutLog;
+    private WorkoutLogEntity workout;
     private List<RunEntity> runs;
     private List<ExerciseEntity> exercises;
 }
