+++ backend/astra_app/src/main/java/com/wenjun/astra_app/service/impl/FitnessServiceImpl.java
@@ -134,7 +134,7 @@ public class FitnessServiceImpl implements FitnessService {
                 .builder()
                 .exercises(exercises)
                 .runs(runs)
-                .workoutLog(workoutLog)
+                .workout(workoutLog)
                 .build();
         return response;
     }
